aws_lambda/download_ecmwf_forecast/requirements.txt
boto3==1.24.2
botocore==1.27.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
ecmwf-opendata==0.1.1
idna==3.3
jmespath==1.0.0
multiurl==0.0.15
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
